requirements.txt
line-bot-sdk
flask
gunicorn
Pillow
psycopg2
urllib3
selenium